Support date separators in virtual message lists
